Updated to match newest commits
I've updated the files and structure to be the same as the newest main commit
Included a Dark Mode option in the settings toggle to use this theme.
Update the en.json to include Dark.
Changed the logo to match the new commit.
Removed the attribute builder and pop outs theming.
Updated sf2e.mjs as it wasn't changing between styles.